Allow  us to work against older FreeIPA server versions
The freeipa-admintools package is only forwards-compatible with
FreeIPA servers.  If we attempt to use this image against an older
FreeIPA server, we will encounter API verson errors when we try to
create a new service in IPA.

This patch eliminates our dependency on the freeipa-admintools
package by adding a new python script that uses the FreeIPA API
directly.  This allows us to specify that the script is operating
at an older API version level, thus giving us the greatest possible
compatibility with FreeIPA server versions.